requirements.txt
pymovements
matplotlib~=3.8.4
polars>=1.28.0
pandas~=2.2.2
numpy~=1.26.4
tqdm~=4.66.4
pillow==10.4.0
fastexcel==0.12.0
# docs
sphinx-book-theme>=1.1.4
myst-nb>=1.3.0
sphinx-copybutton
sphinxcontrib-bibtex>=2.6.5
sphinx-autobuild
sphinx_design>=0.6.1